refactor(backend): Phase 6 - colocate provider stubs with their adapters
Deletes the grab-bag `src/providers/stubs.ts` and colocates each stub class
next to its real Ai-SDK adapter. Addresses the architect-reviewer's
original Phase 1 finding that stubs sitting in a shared file forced every
adapter module's type contract to be imported in one place and made the
stub/real selection implicit.

Stub relocations:
- StubArticleExtractor -> src/providers/extractors/markdown.ts
- StubYouTubeExtractor -> src/providers/extractors/youtube/index.ts
- StubDocumentMetadataLlm -> src/providers/llm/documentMetadataLlm.ts
- StubPostDraftLlm -> src/providers/llm/postDraftLlm.ts
- StubPostDraftValidator -> src/providers/llm/postDraftValidator.ts
- StubSectionDraftRankerLlm -> src/providers/llm/sectionDraftRankerLlm.ts
- StubThematicLlm -> src/providers/llm/thematicLlm.ts
- StubConnectionDiscoveryLlm -> src/providers/llm/connectionDiscoveryLlm.ts
- StubHighlightDraftLlm -> src/providers/llm/highlightDraftLlm.ts

Importer updates:
- src/providers/wiring.ts pulls StubArticleExtractor / StubYouTubeExtractor
  from their adapter files.
- convex/indexing/services.ts pulls StubDocumentMetadataLlm alongside
  AiSdkDocumentMetadataLlm from the adapter file.
- convex/drafting/services.ts pulls each drafting stub from its adapter.
- tests/providers/stubs.test.ts imports StubPostDraftValidator from the
  adapter file.

Intentionally deferred (architect-reviewer original P1 findings):
- Entitlements extraction (resolveTier / computeDocumentUsage /
  enforceDocumentLimit into src/access/logic/) - the logic is ~30 lines
  pre-bound to ctx.db and Polar; introducing a PolarSubscriptionProvider
  port now would be speculative abstraction with no current non-Convex
  caller. Re-evaluate if Billing ever needs to read Entitlements.
- Split src/providers/types.ts per capability - 375 lines of cohesive
  port declarations; splitting would multiply import lines without
  reducing real coupling.

Validation: 290/290 backend tests passing. Convex tsc clean. Frontend tsc
clean. Root backend tsc has only the 3 pre-existing summaryPoints.test.ts
errors unrelated to this phase.
