+++ packages/backend/convex/drafting/services.ts
@@ -1,19 +1,23 @@
 "use node";
 
-import { AiSdkPostDraftLlm } from "../../src/providers/llm/postDraftLlm";
-import { AiSdkPostDraftValidator } from "../../src/providers/llm/postDraftValidator";
-import { AiSdkConnectionDiscoveryLlm } from "../../src/providers/llm/connectionDiscoveryLlm";
-import { AiSdkHighlightDraftLlm } from "../../src/providers/llm/highlightDraftLlm";
-import { AiSdkSectionDraftRankerLlm } from "../../src/providers/llm/sectionDraftRankerLlm";
-import { AiSdkThematicLlm } from "../../src/providers/llm/thematicLlm";
 import {
-  StubPostDraftLlm,
-  StubPostDraftValidator,
+  AiSdkConnectionDiscoveryLlm,
   StubConnectionDiscoveryLlm,
+} from "../../src/providers/llm/connectionDiscoveryLlm";
+import {
+  AiSdkHighlightDraftLlm,
   StubHighlightDraftLlm,
+} from "../../src/providers/llm/highlightDraftLlm";
+import { AiSdkPostDraftLlm, StubPostDraftLlm } from "../../src/providers/llm/postDraftLlm";
+import {
+  AiSdkPostDraftValidator,
+  StubPostDraftValidator,
+} from "../../src/providers/llm/postDraftValidator";
+import {
+  AiSdkSectionDraftRankerLlm,
   StubSectionDraftRankerLlm,
-  StubThematicLlm,
-} from "../../src/providers/stubs";
+} from "../../src/providers/llm/sectionDraftRankerLlm";
+import { AiSdkThematicLlm, StubThematicLlm } from "../../src/providers/llm/thematicLlm";
 import type {
   ConnectionDiscoveryServiceContext,
   DraftGenerationServiceContext,
